requirements.txt
fastapi[standard]
pydantic-settings
uvicorn
python-dotenv
httpx

sqlmodel
fastapi-sso
python-jose[cryptography]
python-multipart
wtforms
pytest